addressed a few of @wlach's comments. Having trouble with revision actually parsing the jsmd in a notebook.
