+++ server/notebooks/urls.py
@@ -1,10 +1,9 @@
 from django.conf.urls import url
-from .views import notebook_view
+from .views import notebook_view, notebook_revisions
 
 
 urlpatterns = [
-    # url(r'^(?P<pk>[0-9]+)/revisions/$', revisions_view,
-    #     name='revisions-view'),
-    url(r'^(?P<pk>[0-9]+)/', notebook_view,
+    url(r'^(?P<pk>[0-9]+)/revisions/', notebook_revisions, name='notebook-revisions'),
+    url(r'^(?P<pk>[0-9]+)/$', notebook_view,
         name='notebook-view'),
 ]
