+++ server/urls.py
@@ -3,16 +3,12 @@ from django.contrib import admin
 from django.urls import path
 
 import server.views
-from .notebooks.views import revision_view
 
 admin.autodiscover()
 
 urlpatterns = [
     # notebook stuff
     url(r'^api/v1/', include('server.notebooks.api_urls')),
-    url(r'^notebooks/(?P<nb>[a-zA-Z0-9]+)/revisions/(?P<pk>[0-9]+)/', revision_view,
-        name='revision-view'),
-    url(r'^notebooks/(?P<pk>[a-zA-Z0-9]+)/revisions', server.views.revisions, name='revisions'),
     url(r'^notebooks/', include('server.notebooks.urls')),
     url(r'^new/?', server.notebooks.views.new_notebook_view,
         name='new-notebook'),
